% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-data-set.r
\name{asUchroniaData}
\alias{asUchroniaData}
\title{Coerce an object to a c++ object handled via an external pointer}
\usage{
asUchroniaData(rData)
}
\arguments{
\item{rData}{An R object such as an xts object.}
}
\value{
an S4 object 'ExternalObjRef' [package "cinterop"]
}
\description{
Converts an object such as an xts time series to an equivalent 'uchronia' C++ entity that it then refered to 
via an external pointer. Typically deals with time series and ensemble thereof, but may be expanded later on to support more types.
}
\examples{
\dontrun{
len <- 3
ensSize <- 2
set.seed(1)
x <- matrix(rnorm(n=len*ensSize), ncol=ensSize)
ind <- lubridate::origin + (1:len) * lubridate::days(1)
(mts <- xts(x, ind))
(umts <- asUchroniaData(mts))
asXts(umts)
}
}
