% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-data-set.r
\name{asXts}
\alias{asXts}
\title{Coerce an object to an xts time series}
\usage{
asXts(tsInfo)
}
\arguments{
\item{tsInfo}{A representation of a time series. Supported types are external pointers as data from uchronia C API, or an R list returned by some of the *_R functions.}
}
\value{
an xts object
}
\description{
Converts if possible an object to an xts time series. Suitable objects are an equivalent 'uchronia' C++ entity 
 via an external pointer. Typically deals with time series and ensemble thereof, but may be expanded later on to support more types.
}
\seealso{
\code{\link{asUchroniaData}} for a round-trip sample code
}
