% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-wrap-generated.r
\name{CreateEnsembleDataset_R}
\alias{CreateEnsembleDataset_R}
\title{CreateEnsembleDataset_R}
\usage{
CreateEnsembleDataset_R(type)
}
\arguments{
\item{type}{R type equivalent for C++ type const char*}
}
\description{
CreateEnsembleDataset_R Wrapper function for CreateEnsembleDataset
}
