% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-time-series.r
\name{createEnsembleForecastTimeSeries}
\alias{createEnsembleForecastTimeSeries}
\title{Create a time series of ensemble of time series}
\usage{
createEnsembleForecastTimeSeries(tsStartEns, n, timeStep = "daily")
}
\arguments{
\item{tsStartEns}{data coercible to a POSIXct start date of the ensemble time series.}

\item{n}{integer. length of the ensemble time series}

\item{timeStep}{a character vector (numeric integers are 'tolerated' as seconds). Examples are '24:00:00', 'hourly', 'daily'.}
}
\description{
Create a time series of ensemble of time series (a.k.a. ensemble forecast time series)
}
\examples{
\dontrun{
n_days <- 10  
tsStartDate <- lubridate::origin
ensFcTs <- createEnsembleForecastTimeSeries(tsStartDate, n_days, 'daily')
fcastsOffset <- lubridate::hours(1)
nLead <- 5
nEns <- 4
set.seed(465)
x <- matrix(rnorm(n=nEns*nLead), ncol=nEns)
for (i in as.integer(1:n_days)) {
  multiTimeSeriesIn <- mkHourlySeries(tsStartDate+ lubridate::hours(i) + fcastsOffset, (x+i-1), is.na)
  uchronia::setItem(ensFcTs, i, multiTimeSeriesIn)
}
print(x)
print(uchronia::getItem(ensFcTs, 1))
}
}
