% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-wrap-generated.r
\name{CreateEnsembleForecastTimeSeries_R}
\alias{CreateEnsembleForecastTimeSeries_R}
\title{CreateEnsembleForecastTimeSeries_R}
\usage{
CreateEnsembleForecastTimeSeries_R(start, length, timeStepName)
}
\arguments{
\item{start}{R type equivalent for C++ type date_time_to_second}

\item{length}{R type equivalent for C++ type int}

\item{timeStepName}{R type equivalent for C++ type const char*}
}
\description{
CreateEnsembleForecastTimeSeries_R Wrapper function for CreateEnsembleForecastTimeSeries
}
