% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-wrap-generated.r
\name{CreateEnsembleTimeSeriesDataFromStruct_R}
\alias{CreateEnsembleTimeSeriesDataFromStruct_R}
\title{CreateEnsembleTimeSeriesDataFromStruct_R}
\usage{
CreateEnsembleTimeSeriesDataFromStruct_R(ensSeries)
}
\arguments{
\item{ensSeries}{R type equivalent for C++ type const multi_regular_time_series_data*}
}
\description{
CreateEnsembleTimeSeriesDataFromStruct_R Wrapper function for CreateEnsembleTimeSeriesDataFromStruct
}
