% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-wrap-generated.r
\name{CreatePerfectForecastTimeSeries_R}
\alias{CreatePerfectForecastTimeSeries_R}
\title{CreatePerfectForecastTimeSeries_R}
\usage{
CreatePerfectForecastTimeSeries_R(observations, start, length, timeStepName,
  offsetForecasts, leadTime)
}
\arguments{
\item{observations}{R type equivalent for C++ type TIME_SERIES_PTR}

\item{start}{R type equivalent for C++ type date_time_to_second}

\item{length}{R type equivalent for C++ type int}

\item{timeStepName}{R type equivalent for C++ type const char*}

\item{offsetForecasts}{R type equivalent for C++ type int}

\item{leadTime}{R type equivalent for C++ type int}
}
\description{
CreatePerfectForecastTimeSeries_R Wrapper function for CreatePerfectForecastTimeSeries
}
