% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-wrap-generated.r
\name{CreateSingleTimeSeriesDataFromStruct_R}
\alias{CreateSingleTimeSeriesDataFromStruct_R}
\title{CreateSingleTimeSeriesDataFromStruct_R}
\usage{
CreateSingleTimeSeriesDataFromStruct_R(timeSeries)
}
\arguments{
\item{timeSeries}{R type equivalent for C++ type const multi_regular_time_series_data*}
}
\description{
CreateSingleTimeSeriesDataFromStruct_R Wrapper function for CreateSingleTimeSeriesDataFromStruct
}
