% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-wrap-generated.r
\name{DeleteDoubleArray_R}
\alias{DeleteDoubleArray_R}
\title{DeleteDoubleArray_R}
\usage{
DeleteDoubleArray_R(value)
}
\arguments{
\item{value}{R type equivalent for C++ type double*}
}
\description{
DeleteDoubleArray_R Wrapper function for DeleteDoubleArray
}
