% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-time-series.r
\name{deserializeSampleSeries}
\alias{deserializeSampleSeries}
\title{Deserialize information to a UTC time series}
\usage{
deserializeSampleSeries(varData)
}
\arguments{
\item{varData}{a list with integer form of UTC POSIXct date stamps ('utcInt'), and series values ('x') as vector or matrix, and column names ('name') where name's value can be NULL.}
}
\value{
an xts time series with UTC time indexing
}
\description{
Deserialize information to a UTC time series. This function is overcoming some behaviors in saving/loading xts series to/from binary RData format. Usage is not intended for most users. See \code{\link{serializableTs}}
}
