% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-wrap-generated.r
\name{DisposeSharedPointer_R}
\alias{DisposeSharedPointer_R}
\title{DisposeSharedPointer_R}
\usage{
DisposeSharedPointer_R(ptr)
}
\arguments{
\item{ptr}{R type equivalent for C++ type VOID_PTR_PROVIDER_PTR}
}
\description{
DisposeSharedPointer_R Wrapper function for DisposeSharedPointer
}
