% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{dotOnLoad}
\alias{dotOnLoad}
\alias{.onLoad}
\title{onLoad function for package}
\usage{
.onLoad(libname = "~/R", pkgname = "uchronia")
}
\arguments{
\item{libname}{the path to the library from which the package is loaded}

\item{pkgname}{the name of the package.}
}
\description{
Function called when loading the package with the function library.
}
