% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-wrap-generated.r
\name{EnsembleSizeEnsembleTimeSeries_R}
\alias{EnsembleSizeEnsembleTimeSeries_R}
\title{EnsembleSizeEnsembleTimeSeries_R}
\usage{
EnsembleSizeEnsembleTimeSeries_R(ensSeries)
}
\arguments{
\item{ensSeries}{R type equivalent for C++ type ENSEMBLE_PTR_TIME_SERIES_PTR}
}
\description{
EnsembleSizeEnsembleTimeSeries_R Wrapper function for EnsembleSizeEnsembleTimeSeries
}
