% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-time-series.r
\name{geometryOf}
\alias{geometryOf}
\title{Gets information about the dimension(s) of data}
\usage{
geometryOf(x)
}
\arguments{
\item{x}{uchronia data such as TIME_SERIES_PTR, 
ENSEMBLE_PTR_TIME_SERIES_PTR, or ENSEMBLE_FORECAST_TIME_SERIES_PTR.}
}
\description{
Gets information about the dimension(s) of data
}
