% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-data-set.r
\name{getDataIdentifiers}
\alias{getDataIdentifiers}
\title{Gets the known time series identifiers (e.g. Gauge names) of a time series provider}
\usage{
getDataIdentifiers(provider)
}
\arguments{
\item{provider}{an external pointer to a uchronia Time Series Provider.}
}
\value{
a character vector
}
\description{
Gets the known time series identifiers (e.g. Gauge names) of a time series provider.
 This means that the argument is a wrapper around an external pointer to a object whose type is 
 inheriting from the C++ type datatypes::timeseries::TimeSeriesProvider<double>
}
