% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-data-set.r
\name{getDataSet}
\alias{getDataSet}
\title{Retrieve data from a data sets library}
\usage{
getDataSet(dataLibrary, dataId)
}
\arguments{
\item{dataLibrary}{R type equivalent for C++ type ENSEMBLE_DATA_SET_PTR, a.k.a a "time series library"}

\item{dataId}{character, one data identifier for the data retrieved.}
}
\value{
R type equivalent for one of the C++ types TIME_SERIES_PTR, 
 ENSEMBLE_PTR_TIME_SERIES_PTR, or ENSEMBLE_FORECAST_TIME_SERIES_PTR.
 A data identifier may thus point to an univariate time series, 
 a multivariate time series, or a time series of ensemble of (uni- or multi- variate) time series.
}
\description{
Gets the data from a library for a given identifier.
}
\seealso{
the vignettes illustrate how to use time serie libraries
}
