% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-wrap-generated.r
\name{GetDatasetEnsembleForecastTimeSeries_R}
\alias{GetDatasetEnsembleForecastTimeSeries_R}
\title{GetDatasetEnsembleForecastTimeSeries_R}
\usage{
GetDatasetEnsembleForecastTimeSeries_R(dataLibrary, dataId)
}
\arguments{
\item{dataLibrary}{R type equivalent for C++ type ENSEMBLE_DATA_SET_PTR}

\item{dataId}{R type equivalent for C++ type const char*}
}
\description{
GetDatasetEnsembleForecastTimeSeries_R Wrapper function for GetDatasetEnsembleForecastTimeSeries
}
