% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-wrap-generated.r
\name{GetDatasetEnsembleTimeSeries_R}
\alias{GetDatasetEnsembleTimeSeries_R}
\title{GetDatasetEnsembleTimeSeries_R}
\usage{
GetDatasetEnsembleTimeSeries_R(dataLibrary, dataEnsembleId)
}
\arguments{
\item{dataLibrary}{R type equivalent for C++ type ENSEMBLE_DATA_SET_PTR}

\item{dataEnsembleId}{R type equivalent for C++ type const char*}
}
\description{
GetDatasetEnsembleTimeSeries_R Wrapper function for GetDatasetEnsembleTimeSeries
}
