% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-pkg-wrap-generated.r
\name{GetDatasetFromLibrary_Pkg_R}
\alias{GetDatasetFromLibrary_Pkg_R}
\title{GetDatasetFromLibrary_Pkg_R}
\usage{
GetDatasetFromLibrary_Pkg_R(dataLibrary, dataIdentifier)
}
\arguments{
\item{dataLibrary}{R type equivalent for C++ type XPtr<opaque_pointer_handle>}

\item{dataIdentifier}{R type equivalent for C++ type CharacterVector}
}
\description{
GetDatasetFromLibrary_Pkg_R Wrapper function for GetDatasetFromLibrary_Pkg
}
