% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-data-set.r
\name{getDataSetIds}
\alias{getDataSetIds}
\title{Gets the top level data identifiers in a data library (data set)}
\usage{
getDataSetIds(dataLibrary)
}
\arguments{
\item{dataLibrary}{R type equivalent for C++ type ENSEMBLE_DATA_SET_PTR, a.k.a a "time series library"}
}
\description{
Gets the top level data identifiers in a data library (data set)
}
