% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-data-set.r
\name{getDatasetSingleTimeSeries}
\alias{getDatasetSingleTimeSeries}
\title{Retrieve data from a data sets library}
\usage{
getDatasetSingleTimeSeries(dataLibrary, dataId)
}
\arguments{
\item{dataLibrary}{R type equivalent for C++ type ENSEMBLE_DATA_SET_PTR}

\item{dataId}{character, one data identifier for the time series.}
}
\description{
Gets the data from a library for a given identifier.
}
