% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-wrap-generated.r
\name{GetDatasetSingleTimeSeries_R}
\alias{GetDatasetSingleTimeSeries_R}
\title{GetDatasetSingleTimeSeries_R}
\usage{
GetDatasetSingleTimeSeries_R(dataLibrary, dataId)
}
\arguments{
\item{dataLibrary}{R type equivalent for C++ type ENSEMBLE_DATA_SET_PTR}

\item{dataId}{R type equivalent for C++ type const char*}
}
\description{
GetDatasetSingleTimeSeries_R Wrapper function for GetDatasetSingleTimeSeries
}
