% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-wrap-generated.r
\name{GetEnsembleDatasetDataIdentifiers_R}
\alias{GetEnsembleDatasetDataIdentifiers_R}
\title{GetEnsembleDatasetDataIdentifiers_R}
\usage{
GetEnsembleDatasetDataIdentifiers_R(dataLibrary)
}
\description{
GetEnsembleDatasetDataIdentifiers_R Wrapper function for GetEnsembleDatasetDataIdentifiers
}
