% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-wrap-generated.r
\name{GetEnsembleDatasetDataSubIdentifiers_R}
\alias{GetEnsembleDatasetDataSubIdentifiers_R}
\title{GetEnsembleDatasetDataSubIdentifiers_R}
\usage{
GetEnsembleDatasetDataSubIdentifiers_R(dataLibrary, dataCollectionId)
}
\description{
GetEnsembleDatasetDataSubIdentifiers_R Wrapper function for GetEnsembleDatasetDataSubIdentifiers
}
