% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-wrap-generated.r
\name{GetEnsembleDatasetDataSummaries_R}
\alias{GetEnsembleDatasetDataSummaries_R}
\title{GetEnsembleDatasetDataSummaries_R}
\usage{
GetEnsembleDatasetDataSummaries_R(dataLibrary)
}
\description{
GetEnsembleDatasetDataSummaries_R Wrapper function for GetEnsembleDatasetDataSummaries
}
