% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-pkg-wrap-generated.r
\name{GetEnsembleForecastTimeSeriesGeometry_Pkg_R}
\alias{GetEnsembleForecastTimeSeriesGeometry_Pkg_R}
\title{GetEnsembleForecastTimeSeriesGeometry_Pkg_R}
\usage{
GetEnsembleForecastTimeSeriesGeometry_Pkg_R(timeSeries)
}
\arguments{
\item{timeSeries}{R type equivalent for C++ type XPtr<opaque_pointer_handle>}
}
\description{
GetEnsembleForecastTimeSeriesGeometry_Pkg_R Wrapper function for GetEnsembleForecastTimeSeriesGeometry_Pkg
}
