% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-wrap-generated.r
\name{GetEnsembleForecastTimeSeriesGeometry_R}
\alias{GetEnsembleForecastTimeSeriesGeometry_R}
\title{GetEnsembleForecastTimeSeriesGeometry_R}
\usage{
GetEnsembleForecastTimeSeriesGeometry_R(timeSeries, geom)
}
\arguments{
\item{timeSeries}{R type equivalent for C++ type ENSEMBLE_FORECAST_TIME_SERIES_PTR}

\item{geom}{R type equivalent for C++ type TS_GEOMETRY_PTR}
}
\description{
GetEnsembleForecastTimeSeriesGeometry_R Wrapper function for GetEnsembleForecastTimeSeriesGeometry
}
