% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-pkg-wrap-generated.r
\name{GetEnsembleTimeSeries_Pkg_R}
\alias{GetEnsembleTimeSeries_Pkg_R}
\title{GetEnsembleTimeSeries_Pkg_R}
\usage{
GetEnsembleTimeSeries_Pkg_R(series)
}
\arguments{
\item{series}{R type equivalent for C++ type XPtr<opaque_pointer_handle>}
}
\description{
GetEnsembleTimeSeries_Pkg_R Wrapper function for GetEnsembleTimeSeries_Pkg
}
