% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-time-series.r
\name{getItem}
\alias{getItem}
\title{Gets an item in an indexable uchronia object}
\usage{
getItem(ensFcTs, i, convertToXts = TRUE)
}
\arguments{
\item{ensFcTs}{R type equivalent for C++ type TIME_SERIES_PTR, 
ENSEMBLE_PTR_TIME_SERIES_PTR, or ENSEMBLE_FORECAST_TIME_SERIES_PTR.}

\item{i}{integer. One-based index in the indexable object.}
}
\description{
Gets an item in an indexable uchronia object of one of the the C++ uchronia types identifed by TIME_SERIES_PTR, 
 ENSEMBLE_PTR_TIME_SERIES_PTR, or ENSEMBLE_FORECAST_TIME_SERIES_PTR.
}
\examples{
\dontrun{
ensFcTs <- createEnsembleForecastTimeSeries(lubridate::origin, 3, 'daily')
fcastsOffset <- lubridate::hours(1)
nLead <- 5
nEns <- 4
set.seed(465)
x <- matrix(rnorm(n=nEns*nLead), ncol=nEns)
multiTimeSeriesIn <- mkHourlySeries(lubridate::origin + fcastsOffset, x, is.na)
uchronia::setItem(ensFcTs, 1, multiTimeSeriesIn)
print(x)
print(uchronia::getItem(ensFcTs, 1))
}
}
