% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-pkg-wrap-generated.r
\name{GetItemEnsembleForecastTimeSeries_Pkg_R}
\alias{GetItemEnsembleForecastTimeSeries_Pkg_R}
\title{GetItemEnsembleForecastTimeSeries_Pkg_R}
\usage{
GetItemEnsembleForecastTimeSeries_Pkg_R(series, i)
}
\arguments{
\item{series}{R type equivalent for C++ type XPtr<opaque_pointer_handle>}

\item{i}{R type equivalent for C++ type IntegerVector}
}
\description{
GetItemEnsembleForecastTimeSeries_Pkg_R Wrapper function for GetItemEnsembleForecastTimeSeries_Pkg
}
