% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-wrap-generated.r
\name{GetItemEnsembleTimeSeriesAsStructure_R}
\alias{GetItemEnsembleTimeSeriesAsStructure_R}
\title{GetItemEnsembleTimeSeriesAsStructure_R}
\usage{
GetItemEnsembleTimeSeriesAsStructure_R(series, i)
}
\arguments{
\item{series}{R type equivalent for C++ type ENSEMBLE_PTR_TIME_SERIES_PTR}

\item{i}{R type equivalent for C++ type int}
}
\description{
GetItemEnsembleTimeSeriesAsStructure_R Wrapper function for GetItemEnsembleTimeSeriesAsStructure
}
