% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-wrap-generated.r
\name{GetLastStdExceptionMessage_R}
\alias{GetLastStdExceptionMessage_R}
\title{GetLastStdExceptionMessage_R}
\usage{
GetLastStdExceptionMessage_R()
}
\description{
GetLastStdExceptionMessage_R Wrapper function for GetLastStdExceptionMessage
}
