% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-data-set.r
\name{getMultipleTimeSeriesFromProvider}
\alias{getMultipleTimeSeriesFromProvider}
\title{Gets one or more time series from a time series provider}
\usage{
getMultipleTimeSeriesFromProvider(tsProvider, varIds, apiGetTsFunc)
}
\arguments{
\item{tsProvider}{R wrapper around an object coercible to a TIME_SERIES_PROVIDER_PTR}

\item{varIds}{character vector, IDs of the time series to retrieve from the provider}

\item{apiGetTsFunc}{an R function that takes as arguments 'tsProvider' and a character.}
}
\value{
an xts time series
}
\description{
Gets one or more time series from a time series provider. This function is exported for use by R packages rather than for end users.
}
\examples{
\dontrun{
internalGetRecordedTts <- function(simulation, varIds) {
  uchronia::getMultipleTimeSeriesFromProvider(simulation, varIds, GetRecorded_Pkg_R)
}
}
}
