% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-wrap-generated.r
\name{GetProviderTimeSeriesIdentifiers_R}
\alias{GetProviderTimeSeriesIdentifiers_R}
\title{GetProviderTimeSeriesIdentifiers_R}
\usage{
GetProviderTimeSeriesIdentifiers_R(dataLibrary)
}
\description{
GetProviderTimeSeriesIdentifiers_R Wrapper function for GetProviderTimeSeriesIdentifiers
}
