% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-wrap-generated.r
\name{GetProviderTimeSeriesValues_R}
\alias{GetProviderTimeSeriesValues_R}
\title{GetProviderTimeSeriesValues_R}
\usage{
GetProviderTimeSeriesValues_R(dataLibrary, variableIdentifier, values,
  arrayLength)
}
\arguments{
\item{dataLibrary}{R type equivalent for C++ type TIME_SERIES_PROVIDER_PTR}

\item{variableIdentifier}{R type equivalent for C++ type const char*}

\item{values}{R type equivalent for C++ type double*}

\item{arrayLength}{R type equivalent for C++ type int}
}
\description{
GetProviderTimeSeriesValues_R Wrapper function for GetProviderTimeSeriesValues
}
