% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-wrap-generated.r
\name{GetProviderTsGeometry_R}
\alias{GetProviderTsGeometry_R}
\title{GetProviderTsGeometry_R}
\usage{
GetProviderTsGeometry_R(dataLibrary, variableIdentifier, geom)
}
\arguments{
\item{dataLibrary}{R type equivalent for C++ type TIME_SERIES_PROVIDER_PTR}

\item{variableIdentifier}{R type equivalent for C++ type const char*}

\item{geom}{R type equivalent for C++ type TS_GEOMETRY_PTR}
}
\description{
GetProviderTsGeometry_R Wrapper function for GetProviderTsGeometry
}
