% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-data-set.r
\name{getTimeSeriesFromProvider}
\alias{getTimeSeriesFromProvider}
\title{Gets a time series from a time provider, given a data ID}
\usage{
getTimeSeriesFromProvider(provider, dataId)
}
\arguments{
\item{provider}{an external pointer to a uchronia Time Series Provider.}

\item{dataId}{character, one or more data identifier for the time series. If missing, all known identifiers will be used; be careful about the resulting size.}
}
\value{
an xts time series.
}
\description{
Gets a time series from a time provider, given a data ID.
 This means that the argument is a wrapper around an external pointer to a object whose type is 
 inheriting from the C++ type datatypes::timeseries::TimeSeriesProvider<double>
}
