% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-pkg-wrap-generated.r
\name{GetTimeSeriesGeometry_Pkg_R}
\alias{GetTimeSeriesGeometry_Pkg_R}
\title{GetTimeSeriesGeometry_Pkg_R}
\usage{
GetTimeSeriesGeometry_Pkg_R(timeSeries)
}
\arguments{
\item{timeSeries}{R type equivalent for C++ type XPtr<opaque_pointer_handle>}
}
\description{
GetTimeSeriesGeometry_Pkg_R Wrapper function for GetTimeSeriesGeometry_Pkg
}
