% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-wrap-generated.r
\name{GetTimeSeriesGeometry_R}
\alias{GetTimeSeriesGeometry_R}
\title{GetTimeSeriesGeometry_R}
\usage{
GetTimeSeriesGeometry_R(timeSeries, geom)
}
\arguments{
\item{timeSeries}{R type equivalent for C++ type TIME_SERIES_PTR}

\item{geom}{R type equivalent for C++ type TS_GEOMETRY_PTR}
}
\description{
GetTimeSeriesGeometry_R Wrapper function for GetTimeSeriesGeometry
}
