% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-wrap-generated.r
\name{GetTimeSeriesValues_R}
\alias{GetTimeSeriesValues_R}
\title{GetTimeSeriesValues_R}
\usage{
GetTimeSeriesValues_R(timeSeries, values, arrayLength)
}
\arguments{
\item{timeSeries}{R type equivalent for C++ type TIME_SERIES_PTR}

\item{values}{R type equivalent for C++ type double*}

\item{arrayLength}{R type equivalent for C++ type int}
}
\description{
GetTimeSeriesValues_R Wrapper function for GetTimeSeriesValues
}
