% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-wrap-generated.r
\name{GetValueFromUnivariateTimeSeries_R}
\alias{GetValueFromUnivariateTimeSeries_R}
\title{GetValueFromUnivariateTimeSeries_R}
\usage{
GetValueFromUnivariateTimeSeries_R(ts, index)
}
\arguments{
\item{ts}{R type equivalent for C++ type TIME_SERIES_PTR}

\item{index}{R type equivalent for C++ type int}
}
\description{
GetValueFromUnivariateTimeSeries_R Wrapper function for GetValueFromUnivariateTimeSeries
}
