% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-time-series.r
\name{hasIdentifier}
\alias{hasIdentifier}
\title{Checks whether a data library has a given top level data identifier}
\usage{
hasIdentifier(dataLibrary, identifier)
}
\arguments{
\item{dataLibrary}{an S4 object 'ExternalObjRef' [package "cinterop"] with external pointer type ENSEMBLE_DATA_SET_PTR}

\item{identifier}{character, the identifier to test again}
}
\description{
Checks whether a data library has a given top level data identifier
}
\seealso{
\code{\link{getEnsembleDataSet}} for sample code
}
