% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-internals.r
\name{internalGetSingleModelTts}
\alias{internalGetSingleModelTts}
\title{Internal only - Perform an action on a uchronia object that is expected to return a time series}
\usage{
internalGetSingleModelTts(tsProvider, varId, apiGetTsFunc)
}
\arguments{
\item{tsProvider}{an external pointer to a uchronia Time Series Provider or model simulation}

\item{varId}{character, a data identifier for the time series}

\item{apiGetTsFunc}{a function, that takes as arguments tsProvider and varId, and returns a list as suitable for \code{\link{marshaledTimeSeriesToXts}}}
}
\value{
an xts time series.
}
\description{
Perform an action on a uchronia object that is expected to return a time series. 
  This function is internal to the package, to prevent code duplication between 
  retrieval of time series from a time series provider, or played/recorded 
  time series into/out of a model simulation
}
