% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-internals.r
\name{isEnsembleTimeSeries}
\alias{isEnsembleTimeSeries}
\title{Is the object a 'uchronia' ensemble of time series}
\usage{
isEnsembleTimeSeries(s)
}
\arguments{
\item{s}{an S4 object 'ExternalObjRef' [package "cinterop"] with external pointer}
}
\description{
Is the object a 'uchronia' ensemble of time series
}
