% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-wrap-generated.r
\name{IsMissingValueItemEnsembleForecastTimeSeries_R}
\alias{IsMissingValueItemEnsembleForecastTimeSeries_R}
\title{IsMissingValueItemEnsembleForecastTimeSeries_R}
\usage{
IsMissingValueItemEnsembleForecastTimeSeries_R(series, i)
}
\arguments{
\item{series}{R type equivalent for C++ type ENSEMBLE_FORECAST_TIME_SERIES_PTR}

\item{i}{R type equivalent for C++ type int}
}
\description{
IsMissingValueItemEnsembleForecastTimeSeries_R Wrapper function for IsMissingValueItemEnsembleForecastTimeSeries
}
