% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-internals.r
\name{isSingularTimeSeries}
\alias{isSingularTimeSeries}
\title{Is the object a 'uchronia' univariate time series}
\usage{
isSingularTimeSeries(s)
}
\arguments{
\item{s}{an S4 object 'ExternalObjRef' [package "cinterop"] with external pointer}
}
\description{
Is the object a 'uchronia' univariate time series
}
