% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-internals.r
\name{isTimeSeriesOfEnsembleTimeSeries}
\alias{isTimeSeriesOfEnsembleTimeSeries}
\title{Is the object a 'uchronia' time series of ensembles of time series}
\usage{
isTimeSeriesOfEnsembleTimeSeries(s)
}
\arguments{
\item{s}{an S4 object 'ExternalObjRef' [package "cinterop"] with external pointer}
}
\description{
Is the object a 'uchronia' time series of ensembles of time series
}
