% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-wrap-generated.r
\name{LoadEnsembleDataset_R}
\alias{LoadEnsembleDataset_R}
\title{LoadEnsembleDataset_R}
\usage{
LoadEnsembleDataset_R(libraryIdentifier, dataPath)
}
\arguments{
\item{libraryIdentifier}{R type equivalent for C++ type const char*}

\item{dataPath}{R type equivalent for C++ type const char*}
}
\description{
LoadEnsembleDataset_R Wrapper function for LoadEnsembleDataset
}
