% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-time-series.r
\name{mkDailySeries}
\alias{mkDailySeries}
\title{Make an daily time series}
\usage{
mkDailySeries(startDate, x, isMissingFunc = function(val) {     val < 0 })
}
\arguments{
\item{startDate}{the start date of the time series. It must be a POSIXct object.}

\item{x}{the numeric vector.}

\item{isMissingFunc}{a function, or NULL. The function should return a logical vector. Used to detect missing values in the raw data that are not already identified as NA}
}
\value{
an xts time series.
}
\description{
Make an daily time series from a numeric vector and a start date.
}
\examples{
\dontrun{
mkDailySeries(lubridate::origin, -2:3/3.0)
}
}
