% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-time-series.r
\name{mkDate}
\alias{mkDate}
\title{Creates a POSIXct date/time object}
\usage{
mkDate(year, month, day, hour = 0, min = 0, sec = 0, tz = "UTC")
}
\arguments{
\item{year}{integer}

\item{month}{integer}

\item{day}{integer}

\item{hour}{integer}

\item{min}{integer}

\item{sec}{numeric}

\item{tz}{character, A time zone specification to be used for the conversion. Defaults to UTC.}
}
\value{
a POSIXct date/time object
}
\description{
Creates a POSIXct date/time object, with a default UTC time zone and zeroes as default time arguments (i.e. midnight)
}
