% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-time-series.r
\name{mkSeriesMonthlyTstep}
\alias{mkSeriesMonthlyTstep}
\title{Make a time series with a monthly step}
\usage{
mkSeriesMonthlyTstep(
  startDate,
  x,
  isMissingFunc = function(val) {     val < 0 }
)
}
\arguments{
\item{startDate}{the start date of the time series. It must be a POSIXct object.}

\item{x}{the numeric vector.}

\item{isMissingFunc}{a function, or NULL. The function should return a logical vector. Used to detect missing values in the raw data that are not already identified as NA}
}
\value{
an xts time series.
}
\description{
Make a time series with a monthly step
}
\examples{
\dontrun{
# make a time series with a time step of 10 minutes
mkSeriesMonthlyTstep(lubridate::origin, -2:3/3.0)
}
}
