% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-time-series.r
\name{mkSeriesRegularTstep}
\alias{mkSeriesRegularTstep}
\title{Make a time series with a regular time step}
\usage{
mkSeriesRegularTstep(
  startDate,
  x,
  isMissingFunc = function(val) {     val < 0 },
  deltaTSec = 3600
)
}
\arguments{
\item{startDate}{the start date of the time series. It must be a POSIXct object.}

\item{x}{the numeric vector.}

\item{isMissingFunc}{a function, or NULL. The function should return a logical vector. Used to detect missing values in the raw data that are not already identified as NA}

\item{deltaTSec}{the time step duration in seconds}
}
\value{
an xts time series.
}
\description{
Make a time series with a regular time step, meaning with the same duration between all steps
}
\examples{
\dontrun{
# make a time series with a time step of 10 minutes
mkSeriesRegularTstep(lubridate::origin, -2:3/3.0, deltaTSec=600)
}
}
