% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-plot.r
\name{plotXtsQuantiles}
\alias{plotXtsQuantiles}
\title{Create a ggplot object of quantiles of a multivariate xts time series}
\usage{
plotXtsQuantiles(tSeries, probs, title, xlabel = "Time", ylabel)
}
\arguments{
\item{tSeries}{a time series of class 'xts'}

\item{probs}{data frame with colnames 'low', 'high'. You should order from wider to narrower confidence intervals, e.g. data.frame(low=c(.05, .25), high=c(0.95, 0.75))}

\item{title}{optional, title for the graph}

\item{xlabel}{x-label}

\item{ylabel}{optional, label for y axis}
}
\description{
Create a ggplot object of quantiles of a multivariate xts time series
}
\examples{
\dontrun{
n <- 15
myrand <- function(i) { i + 0.5*sqrt(i) * rnorm(100)}
set.seed(0)
d <- t(sapply(seq_len(n), myrand))
ensTs <- xts::xts(d, order.by=lubridate::origin + 0:(n-1) * 3600)
g <- plotXtsQuantiles(ensTs, probs = data.frame(low=c(.05, .1, .25), high=c(0.95, .9, 0.75)))
g
# To customize your color scheme you can use scale_fill_brewer
g + ggplot2::scale_fill_brewer(palette="Reds")
}
}
