% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-pkg-wrap-generated.r
\name{RegisterExceptionCallback_Pkg_R}
\alias{RegisterExceptionCallback_Pkg_R}
\title{RegisterExceptionCallback_Pkg_R}
\usage{
RegisterExceptionCallback_Pkg_R()
}
\description{
RegisterExceptionCallback_Pkg_R Wrapper function for RegisterExceptionCallback_Pkg
}
