% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-wrap-generated.r
\name{RegisterExceptionCallback_R}
\alias{RegisterExceptionCallback_R}
\title{RegisterExceptionCallback_R}
\usage{
RegisterExceptionCallback_R(callback)
}
\arguments{
\item{callback}{R type equivalent for C++ type const void*}
}
\description{
RegisterExceptionCallback_R Wrapper function for RegisterExceptionCallback
}
