% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-sample-data.r
\name{sampleDataDir}
\alias{sampleDataDir}
\title{Probe for the location of sample data directory (too big for package inclusion)}
\usage{
sampleDataDir(doWarn = TRUE)
}
\arguments{
\item{doWarn}{issue a warning if SWIFT_SAMPLE_DATA_DIR env var is not set}
}
\value{
a string
}
\description{
Probe for the location of sample data directory (too big for package inclusion), defined by the SWIFT_SAMPLE_DATA_DIR environment variable.
}
