% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-sample-data.r
\name{sampleTimeSeriesLibrary}
\alias{sampleTimeSeriesLibrary}
\title{Get one of the sample time series libraries included in this package}
\usage{
sampleTimeSeriesLibrary(ident)
}
\arguments{
\item{ident}{an identifier for the library. Choice (case insensitive): 'Upper Murray', 'Stanley'}
}
\value{
a string
}
\description{
Get one of the sample time series libraries included in this package
}
