% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-wrap-generated.r
\name{SaveEnsembleForecastTimeSeriesToNetcdf_R}
\alias{SaveEnsembleForecastTimeSeriesToNetcdf_R}
\title{SaveEnsembleForecastTimeSeriesToNetcdf_R}
\usage{
SaveEnsembleForecastTimeSeriesToNetcdf_R(tsEnsTs, filename, overwrite)
}
\arguments{
\item{tsEnsTs}{R type equivalent for C++ type ENSEMBLE_FORECAST_TIME_SERIES_PTR}

\item{filename}{R type equivalent for C++ type const char*}

\item{overwrite}{R type equivalent for C++ type bool}
}
\description{
SaveEnsembleForecastTimeSeriesToNetcdf_R Wrapper function for SaveEnsembleForecastTimeSeriesToNetcdf
}
