% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-wrap-generated.r
\name{SaveEnsembleTimeSeriesToNetcdf_R}
\alias{SaveEnsembleTimeSeriesToNetcdf_R}
\title{SaveEnsembleTimeSeriesToNetcdf_R}
\usage{
SaveEnsembleTimeSeriesToNetcdf_R(collection, filename, overwrite)
}
\arguments{
\item{collection}{R type equivalent for C++ type ENSEMBLE_PTR_TIME_SERIES_PTR}

\item{filename}{R type equivalent for C++ type const char*}

\item{overwrite}{R type equivalent for C++ type bool}
}
\description{
SaveEnsembleTimeSeriesToNetcdf_R Wrapper function for SaveEnsembleTimeSeriesToNetcdf
}
