% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-wrap-generated.r
\name{SaveSingleTimeSeriesToNetcdf_R}
\alias{SaveSingleTimeSeriesToNetcdf_R}
\title{SaveSingleTimeSeriesToNetcdf_R}
\usage{
SaveSingleTimeSeriesToNetcdf_R(timeSeries, filename, overwrite)
}
\arguments{
\item{timeSeries}{R type equivalent for C++ type TIME_SERIES_PTR}

\item{filename}{R type equivalent for C++ type const char*}

\item{overwrite}{R type equivalent for C++ type bool}
}
\description{
SaveSingleTimeSeriesToNetcdf_R Wrapper function for SaveSingleTimeSeriesToNetcdf
}
