% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-time-series.r
\name{serializableTs}
\alias{serializableTs}
\title{Make UTC time series in a serializable form}
\usage{
serializableTs(series)
}
\arguments{
\item{series}{an xts time series with time indexing in UTC}
}
\value{
a list with integer form of UTC POSIXct date stamps ('utcInt'), series values ('x') as vector or matrix, and column names ('name') where name's value can be NULL.
}
\description{
Make UTC time series in a serializable form. Serializing xts objects directly leads to issues.
}
\examples{
\dontrun{
# make a time series with a time step of 10 minutes
x <- mkSeriesRegularTstep(lubridate::origin, -2:3/3.0, deltaTSec=600)
serializableTs(x)
}
}
