% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-time-series.r
\name{setItem}
\alias{setItem}
\title{Sets an item of an ensemble time series}
\usage{
setItem(ensFcTs, i, value)
}
\arguments{
\item{ensFcTs}{R type equivalent for C++ type ENSEMBLE_FORECAST_TIME_SERIES_PTR}

\item{i}{integer. One-based index in the ensemble forecast.}

\item{value}{an xts object to put the the specified index}
}
\description{
Sets an item of an ensemble time series
}
\examples{
\dontrun{
ensFcTs <- createEnsembleForecastTimeSeries(lubridate::origin, 3, 'daily')
fcastsOffset <- lubridate::hours(1)
nLead <- 5
nEns <- 4
set.seed(465)
x <- matrix(rnorm(n=nEns*nLead), ncol=nEns)
multiTimeSeriesIn <- mkHourlySeries(lubridate::origin + fcastsOffset, x, is.na)
uchronia::setItem(ensFcTs, 1, multiTimeSeriesIn)
print(x)
print(uchronia::getItem(ensFcTs, 1))
}
}
