% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-pkg-wrap-generated.r
\name{SetItemEnsembleForecastTimeSeries_Pkg_R}
\alias{SetItemEnsembleForecastTimeSeries_Pkg_R}
\title{SetItemEnsembleForecastTimeSeries_Pkg_R}
\usage{
SetItemEnsembleForecastTimeSeries_Pkg_R(series, i, timeSeriesEnsemble)
}
\arguments{
\item{series}{R type equivalent for C++ type XPtr<opaque_pointer_handle>}

\item{i}{R type equivalent for C++ type IntegerVector}

\item{timeSeriesEnsemble}{R type equivalent for C++ type Rcpp::S4&}
}
\description{
SetItemEnsembleForecastTimeSeries_Pkg_R Wrapper function for SetItemEnsembleForecastTimeSeries_Pkg
}
