% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-wrap-generated.r
\name{SetItemEnsembleTimeSeriesAsStructure_R}
\alias{SetItemEnsembleTimeSeriesAsStructure_R}
\title{SetItemEnsembleTimeSeriesAsStructure_R}
\usage{
SetItemEnsembleTimeSeriesAsStructure_R(collection, i, values)
}
\arguments{
\item{collection}{R type equivalent for C++ type ENSEMBLE_PTR_TIME_SERIES_PTR}

\item{i}{R type equivalent for C++ type int}

\item{values}{R type equivalent for C++ type const multi_regular_time_series_data*}
}
\description{
SetItemEnsembleTimeSeriesAsStructure_R Wrapper function for SetItemEnsembleTimeSeriesAsStructure
}
