% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-wrap-generated.r
\name{SetTimeSeriesMissingValueValue_R}
\alias{SetTimeSeriesMissingValueValue_R}
\title{SetTimeSeriesMissingValueValue_R}
\usage{
SetTimeSeriesMissingValueValue_R(missingValueValue)
}
\arguments{
\item{missingValueValue}{R type equivalent for C++ type double}
}
\description{
SetTimeSeriesMissingValueValue_R Wrapper function for SetTimeSeriesMissingValueValue
}
