% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-wrap-generated.r
\name{SetValueToUnivariateTimeSeries_R}
\alias{SetValueToUnivariateTimeSeries_R}
\title{SetValueToUnivariateTimeSeries_R}
\usage{
SetValueToUnivariateTimeSeries_R(ts, index, value)
}
\arguments{
\item{ts}{R type equivalent for C++ type TIME_SERIES_PTR}

\item{index}{R type equivalent for C++ type int}

\item{value}{R type equivalent for C++ type double}
}
\description{
SetValueToUnivariateTimeSeries_R Wrapper function for SetValueToUnivariateTimeSeries
}
