% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-time-series.r
\name{subIdentifiers}
\alias{subIdentifiers}
\title{Gets the next level data identifier of a top level ID}
\usage{
subIdentifiers(dataLibrary, identifier)
}
\arguments{
\item{dataLibrary}{an S4 object 'ExternalObjRef' [package "cinterop"] with external pointer type ENSEMBLE_DATA_SET_PTR}

\item{identifier}{character, the top level identifier to test again for next level ids}
}
\description{
Gets the next level data identifier of a top level ID.
 A collection of time series such as one identified by "streamflows" may have 
 sub-identifiers such as gauge numbers. A single time series in a data library 
 may thus be retrieved by a hierarchical string ID  "streamflows.401221" 401221 is a gauge ID.
}
\seealso{
\code{\link{getEnsembleDataSet}} for sample code
}
