% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-wrap-generated.r
\name{TimeSeriesFromEnsembleOfTimeSeries_R}
\alias{TimeSeriesFromEnsembleOfTimeSeries_R}
\title{TimeSeriesFromEnsembleOfTimeSeries_R}
\usage{
TimeSeriesFromEnsembleOfTimeSeries_R(collectionTs, index)
}
\arguments{
\item{collectionTs}{R type equivalent for C++ type ENSEMBLE_PTR_TIME_SERIES_PTR}

\item{index}{R type equivalent for C++ type int}
}
\description{
TimeSeriesFromEnsembleOfTimeSeries_R Wrapper function for TimeSeriesFromEnsembleOfTimeSeries
}
