% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-wrap-generated.r
\name{TimeSeriesFromProviderTs_R}
\alias{TimeSeriesFromProviderTs_R}
\title{TimeSeriesFromProviderTs_R}
\usage{
TimeSeriesFromProviderTs_R(dataLibrary, variableIdentifier)
}
\arguments{
\item{dataLibrary}{R type equivalent for C++ type TIME_SERIES_PROVIDER_PTR}

\item{variableIdentifier}{R type equivalent for C++ type const char*}
}
\description{
TimeSeriesFromProviderTs_R Wrapper function for TimeSeriesFromProviderTs
}
