% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-wrap-generated.r
\name{TimeSeriesFromTimeSeriesOfEnsembleOfTimeSeries_R}
\alias{TimeSeriesFromTimeSeriesOfEnsembleOfTimeSeries_R}
\title{TimeSeriesFromTimeSeriesOfEnsembleOfTimeSeries_R}
\usage{
TimeSeriesFromTimeSeriesOfEnsembleOfTimeSeries_R(efts, indexInIssueTime,
  indexInForecastTime)
}
\arguments{
\item{efts}{R type equivalent for C++ type ENSEMBLE_FORECAST_TIME_SERIES_PTR}

\item{indexInIssueTime}{R type equivalent for C++ type int}

\item{indexInForecastTime}{R type equivalent for C++ type int}
}
\description{
TimeSeriesFromTimeSeriesOfEnsembleOfTimeSeries_R Wrapper function for TimeSeriesFromTimeSeriesOfEnsembleOfTimeSeries
}
