% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-pkg-wrap-generated.r
\name{TimeSeriesToTsInfo_Pkg_R}
\alias{TimeSeriesToTsInfo_Pkg_R}
\title{TimeSeriesToTsInfo_Pkg_R}
\usage{
TimeSeriesToTsInfo_Pkg_R(timeSeries)
}
\arguments{
\item{timeSeries}{R type equivalent for C++ type XPtr<opaque_pointer_handle>}
}
\description{
TimeSeriesToTsInfo_Pkg_R Wrapper function for TimeSeriesToTsInfo_Pkg
}
