% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-wrap-generated.r
\name{ToStructEnsembleTimeSeriesData_R}
\alias{ToStructEnsembleTimeSeriesData_R}
\title{ToStructEnsembleTimeSeriesData_R}
\usage{
ToStructEnsembleTimeSeriesData_R(ensSeries)
}
\arguments{
\item{ensSeries}{R type equivalent for C++ type ENSEMBLE_PTR_TIME_SERIES_PTR}
}
\description{
ToStructEnsembleTimeSeriesData_R Wrapper function for ToStructEnsembleTimeSeriesData
}
