% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-wrap-generated.r
\name{ToStructSingleTimeSeriesData_R}
\alias{ToStructSingleTimeSeriesData_R}
\title{ToStructSingleTimeSeriesData_R}
\usage{
ToStructSingleTimeSeriesData_R(timeSeries)
}
\arguments{
\item{timeSeries}{R type equivalent for C++ type TIME_SERIES_PTR}
}
\description{
ToStructSingleTimeSeriesData_R Wrapper function for ToStructSingleTimeSeriesData
}
