% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-wrap-generated.r
\name{TransformEachItem_R}
\alias{TransformEachItem_R}
\title{TransformEachItem_R}
\usage{
TransformEachItem_R(tsEnsTs, method, methodArgument)
}
\arguments{
\item{tsEnsTs}{R type equivalent for C++ type ENSEMBLE_FORECAST_TIME_SERIES_PTR}

\item{method}{R type equivalent for C++ type char*}

\item{methodArgument}{R type equivalent for C++ type char*}
}
\description{
TransformEachItem_R Wrapper function for TransformEachItem
}
