% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchronia-time-series-transforms.r
\name{xtsApply}
\alias{xtsApply}
\title{Apply a function to each row of a time series}
\usage{
xtsApply(
  tSeries,
  func = function(x) {     quantile(x, probs = c(0.05, 0.5, 0.95)) }
)
}
\arguments{
\item{tSeries}{a time series of class 'xts'}

\item{func}{a function to apply to time series values for each time stamp.}
}
\description{
Apply a function to each row of a time series
}
\examples{
\dontrun{
n <- 15
myrand <- function(i) { i + 0.5*sqrt(i) * rnorm(100)}
set.seed(0)
d <- laply(seq_len(n), myrand)
ensTs <- xts(d, order.by=lubridate::origin + 0:(n-1) * 3600)
xtsApply(ensTs)
}
}
